fix(hide_reputation): the wrong field was hidden
Mbin added more fields, like a cake day and the software used by remote users. The mod was selecting the correct field using its number, which was offset by these new ones being placed before it. I changed it to instead select based on the link, which should be more robust even with new ones added in the future.